requirements.txt
cos-python-sdk-v5>=1.9.24
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.3.0
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=1.0.0